feat: support right semi and revert left mark
