+++ src/query/service/src/pipelines/processors/transforms/hash_join/probe_join/right_semi_anti_join.rs
@@ -18,6 +18,7 @@ use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::DataBlock;
 use databend_common_expression::Expr;
+use databend_common_expression::FilterExecutor;
 use databend_common_expression::KeyAccessor;
 use databend_common_hashtable::HashJoinHashtableLike;
 use databend_common_hashtable::RowPtr;
@@ -136,12 +137,7 @@ impl HashJoinProbeState {
         // Build states.
         let build_state = unsafe { &mut *self.hash_join_state.build_state.get() };
         let outer_scan_map = &mut build_state.outer_scan_map;
-        let other_predicate = self
-            .hash_join_state
-            .hash_join_desc
-            .other_predicate
-            .as_ref()
-            .unwrap();
+        let filter_executor = probe_state.filter_executor.as_mut().unwrap();
 
         // Results.
         let mut matched_idx = 0;
@@ -175,7 +171,7 @@ impl HashJoinProbeState {
                         &mut probe_state.generation_state,
                         &build_state.generation_state,
                         outer_scan_map,
-                        other_predicate,
+                        filter_executor,
                     )?;
                     (matched_idx, incomplete_ptr) = self.fill_probe_and_build_indexes::<_, false>(
                         hash_table,
@@ -212,7 +208,7 @@ impl HashJoinProbeState {
                         &mut probe_state.generation_state,
                         &build_state.generation_state,
                         outer_scan_map,
-                        other_predicate,
+                        filter_executor,
                     )?;
                     (matched_idx, incomplete_ptr) = self.fill_probe_and_build_indexes::<_, false>(
                         hash_table,
@@ -236,7 +232,7 @@ impl HashJoinProbeState {
                 &mut probe_state.generation_state,
                 &build_state.generation_state,
                 outer_scan_map,
-                other_predicate,
+                filter_executor,
             )?;
         }
 
@@ -278,7 +274,7 @@ impl HashJoinProbeState {
         probe_state: &mut ProbeBlockGenerationState,
         build_state: &BuildBlockGenerationState,
         outer_scan_map: &mut [Vec<bool>],
-        other_predicate: &Expr,
+        filter_executor: &mut FilterExecutor,
     ) -> Result<()> {
         if self.hash_join_state.interrupt.load(Ordering::Relaxed) {
             return Err(ErrorCode::AbortedQuery(
@@ -310,8 +306,8 @@ impl HashJoinProbeState {
         let result_block = self.merge_eq_block(probe_block, build_block, matched_idx);
 
         if !result_block.is_empty() {
-            let (bm, all_true, all_false) =
-                self.get_other_filters(&result_block, other_predicate, &self.func_ctx)?;
+            let (selection, all_true, all_false) =
+                self.get_other_predicate_selection(filter_executor, &result_block)?;
 
             if all_true {
                 for row_ptr in &build_indexes[0..matched_idx] {
@@ -322,20 +318,13 @@ impl HashJoinProbeState {
                     };
                 }
             } else if !all_false {
-                // Safe to unwrap.
-                let validity = bm.unwrap();
-                let mut idx = 0;
-                while idx < matched_idx {
+                for idx in selection {
                     unsafe {
-                        let valid = validity.get_bit_unchecked(idx);
-                        if valid {
-                            let row_ptr = build_indexes.get_unchecked(idx);
-                            *outer_scan_map
-                                .get_unchecked_mut(row_ptr.chunk_index as usize)
-                                .get_unchecked_mut(row_ptr.row_index as usize) = true;
-                        }
+                        let row_ptr = build_indexes.get_unchecked(*idx as usize);
+                        *outer_scan_map
+                            .get_unchecked_mut(row_ptr.chunk_index as usize)
+                            .get_unchecked_mut(row_ptr.row_index as usize) = true;
                     }
-                    idx += 1;
                 }
             }
         }
